+++ DynamicNightTime/Patches/GameClockPatch.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Microsoft.Xna.Framework;
+using StardewModdingAPI;
 using StardewValley;
 using TwilightShards.Stardew.Common;
 
@@ -16,6 +17,7 @@ namespace DynamicNightTime.Patches
             sunsetT.AddTime(-20);
             
             int astronTime = DynamicNightTime.GetMorningAstroTwilight().ReturnIntTime();
+            DynamicNightTime.Logger.Log($"Latitude for calc is {DynamicNightTime.GetCurrentLocationLat()}", LogLevel.Info);
 
             //colors
             //sunrise 255,159,80