+++ DynamicNightTime/Patches/GetFullyDarkPatch.cs
@@ -1,7 +1,4 @@
-﻿using StardewModdingAPI.Utilities;
-using System;
-using TwilightShards.Common;
-using TwilightShards.Stardew.Common;
+﻿using TwilightShards.Stardew.Common;
 
 namespace DynamicNightTime.Patches
 {
@@ -9,10 +6,10 @@ namespace DynamicNightTime.Patches
     {
         public static void Postfix(ref int __result)
         {
-                SDVTime calcTime = DynamicNightTime.GetNavalTwilight();
-                calcTime.ClampToTenMinutes();
+            SDVTime calcTime = DynamicNightTime.GetNavalTwilight();
+            calcTime.ClampToTenMinutes();
 
-                __result = calcTime.ReturnIntTime();
+            __result = calcTime.ReturnIntTime();
         }
     }
 }